+++ packages/kit/src/views/Prime/hooks/usePrimePaymentMethods.native.ts
@@ -162,7 +162,7 @@ export function usePrimePaymentMethods(): IUsePrimePayment {
           .toNumber();
       }
 
-      const unit =
+      const currency =
         primePaymentUtils.extractCurrencySymbol(priceString, {
           useShortUSSymbol: true,
         }) ||
@@ -176,17 +176,19 @@ export function usePrimePaymentMethods(): IUsePrimePayment {
       packages.push({
         subscriptionPeriod: subscriptionPeriod as ISubscriptionPeriod,
         pricePerYear: pricePerYear || 0,
-        pricePerYearString: `${unit}${new BigNumber(pricePerYear || 0).toFixed(
-          2,
-        )}`,
+        pricePerYearString: `${currency}${new BigNumber(
+          pricePerYear || 0,
+        ).toFixed(2)}`,
         pricePerMonth: pricePerMonth || 0,
-        pricePerMonthString: `${unit}${new BigNumber(
+        pricePerMonthString: `${currency}${new BigNumber(
           pricePerMonth || 0,
         ).toFixed(2)}`,
         priceTotalPerYearString:
           subscriptionPeriod === 'P1M'
-            ? `${unit}${new BigNumber(pricePerMonth || 0).times(12).toFixed(2)}`
-            : `${unit}${new BigNumber(pricePerYear || 0).toFixed(2)}`,
+            ? `${currency}${new BigNumber(pricePerMonth || 0)
+                .times(12)
+                .toFixed(2)}`
+            : `${currency}${new BigNumber(pricePerYear || 0).toFixed(2)}`,
       });
     });
